requirements.txt
torch>=2.0.0
torchvision
gymnasium[box2d]
numpy
opencv-python
tqdm
tensorboard
pyyaml